requirements.txt
pillow
sympy
numpy
boto3